Match protocol relative URLs (`//`)
Protocol-relative links (PRL), also known as protocol-relative URLs (PRURL), are URLs that have no protocol specified. For example, `//example.com` will use the protocol of the current page, typically HTTP or HTTPS.

These URLs commonly occur in the wild, specifically on pages that support both http and https protocols.

Since the URLs without a protocol belong to the class of relative URLs,
it is up to the user to decide what protocol to use when parsing these
URLs.
Example configuration for the url crate:
https://docs.rs/url/2.2.1/url/#base-url