Add support for skipping wildcard URLs
URLs can contain wildcards as in http://*.example.org.
These are commonly used in cases where listing all possible URLs would
not be feasible, e.g. when defining Content-Security-Policy rules [1].

It can be surprising to users of linkify that these URLs get extracted.
As such, this change adds support for disabling the extraction of wildcard
URLs.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy