Remove no longer necessary check if domain is actually an email address
The check was necessary before because we didn't check host names
properly. Now that we reject a `@` as part of a hostname, it won't be
recognized as a plain domain link anymore, so we don't need to check for
emails.

Makes the code nicer and faster - we'd do the email scan for every `.`
trigger before.